Contact entity refactor

+++ src/IBLTermocasa.Blazor/Pages/Crm/Contacts.razor.cs
@@ -1,26 +1,25 @@
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Globalization;
-using System.Web;
 using Blazorise;
 using Blazorise.DataGrid;
-using Volo.Abp.BlazoriseUI.Components;
-using Microsoft.AspNetCore.Authorization;
-using Volo.Abp.Application.Dtos;
-using Volo.Abp.AspNetCore.Components.Web.Theming.PageToolbars;
+using IBLTermocasa.Blazor.Components;
 using IBLTermocasa.Contacts;
 using IBLTermocasa.Permissions;
-using IBLTermocasa.Shared;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Components;
+using Volo.Abp.Application.Dtos;
+using Volo.Abp.AspNetCore.Components.Web.Theming.PageToolbars;
+using Volo.Abp.BlazoriseUI.Components;
+using BreadcrumbItem = Volo.Abp.BlazoriseUI.BreadcrumbItem;
 
-
-
-namespace IBLTermocasa.Blazor.Pages
+namespace IBLTermocasa.Blazor.Pages.Crm
 {
     public partial class Contacts
     {
-        protected List<Volo.Abp.BlazoriseUI.BreadcrumbItem> BreadcrumbItems = new List<Volo.Abp.BlazoriseUI.BreadcrumbItem>();
+        protected List<BreadcrumbItem> BreadcrumbItems = new List<BreadcrumbItem>();
         protected PageToolbar Toolbar {get;} = new PageToolbar();
         protected bool ShowAdvancedFilters { get; set; }
         private IReadOnlyList<ContactDto> ContactList { get; set; }
@@ -31,28 +30,18 @@ namespace IBLTermocasa.Blazor.Pages
         private bool CanCreateContact { get; set; }
         private bool CanEditContact { get; set; }
         private bool CanDeleteContact { get; set; }
-        private ContactCreateDto NewContact { get; set; }
-        private Validations NewContactValidations { get; set; } = new();
-        private ContactUpdateDto EditingContact { get; set; }
-        private Validations EditingContactValidations { get; set; } = new();
-        private Guid EditingContactId { get; set; }
-        private Modal CreateContactModal { get; set; } = new();
-        private Modal EditContactModal { get; set; } = new();
+        private ContactDto ContactInput { get; set; }
         private GetContactsInput Filter { get; set; }
         private DataGridEntityActionsColumn<ContactDto> EntityActionsColumn { get; set; } = new();
         protected string SelectedCreateTab = "contact-create-tab";
         protected string SelectedEditTab = "contact-edit-tab";
         private ContactDto? SelectedContact;
-        
-        
-        
-        
-        
+        [Inject]
+        private SecureConfirmationService _SecureConfirmationService { get; set; }
         
         public Contacts()
         {
-            NewContact = new ContactCreateDto();
-            EditingContact = new ContactUpdateDto();
+            ContactInput = new ContactDto();
             Filter = new GetContactsInput
             {
                 MaxResultCount = PageSize,
@@ -60,14 +49,11 @@ namespace IBLTermocasa.Blazor.Pages
                 Sorting = CurrentSorting
             };
             ContactList = new List<ContactDto>();
-            
-            
         }
 
         protected override async Task OnInitializedAsync()
         {
             await SetPermissionsAsync();
-            
         }
 
         protected override async Task OnAfterRenderAsync(bool firstRender)
@@ -82,7 +68,7 @@ namespace IBLTermocasa.Blazor.Pages
 
         protected virtual ValueTask SetBreadcrumbItemsAsync()
         {
-            BreadcrumbItems.Add(new Volo.Abp.BlazoriseUI.BreadcrumbItem(L["Menu:Contacts"]));
+            BreadcrumbItems.Add(new BreadcrumbItem(L["Menu:Contacts"]));
             return ValueTask.CompletedTask;
         }
 
@@ -106,8 +92,6 @@ namespace IBLTermocasa.Blazor.Pages
                             .IsGrantedAsync(IBLTermocasaPermissions.Contacts.Edit);
             CanDeleteContact = await AuthorizationService
                             .IsGrantedAsync(IBLTermocasaPermissions.Contacts.Delete);
-                            
-                            
         }
 
         private async Task GetContactsAsync()
@@ -119,8 +103,6 @@ namespace IBLTermocasa.Blazor.Pages
             var result = await ContactsAppService.GetListAsync(Filter);
             ContactList = result.Items;
             TotalCount = (int)result.TotalCount;
-            
-            
         }
 
         protected virtual async Task SearchAsync()
@@ -140,7 +122,7 @@ namespace IBLTermocasa.Blazor.Pages
                 culture = "&culture=" + culture;
             }
             await RemoteServiceConfigurationProvider.GetConfigurationOrDefaultOrNullAsync("Default");
-            NavigationManager.NavigateTo($"{remoteService?.BaseUrl.EnsureEndsWith('/') ?? string.Empty}api/app/contacts/as-excel-file?DownloadToken={token}&FilterText={HttpUtility.UrlEncode(Filter.FilterText)}{culture}&Title={HttpUtility.UrlEncode(Filter.Title)}&Name={HttpUtility.UrlEncode(Filter.Name)}&Surname={HttpUtility.UrlEncode(Filter.Surname)}&ConfidentialName={HttpUtility.UrlEncode(Filter.ConfidentialName)}&JobRole={HttpUtility.UrlEncode(Filter.JobRole)}&MailInfo={HttpUtility.UrlEncode(Filter.MailInfo)}&PhoneInfo={HttpUtility.UrlEncode(Filter.PhoneInfo)}&AddressInfo={HttpUtility.UrlEncode(Filter.AddressInfo)}&Tag={HttpUtility.UrlEncode(Filter.Tag)}", forceLoad: true);
+            NavigationManager.NavigateTo($"{remoteService?.BaseUrl.EnsureEndsWith('/') ?? string.Empty}api/app/contacts/as-excel-file?DownloadToken={token}&FilterText={Filter.FilterText}{culture}&Title={Filter.Title}&Name={Filter.Name}&Surname={Filter.Surname}&Phone={Filter.PhoneInfo}&Mail={Filter.MailInfo}&ConfidentialName={Filter.ConfidentialName}&JobRole={Filter.JobRole}&Tag={Filter.Tag}", forceLoad: true);
         }
 
         private async Task OnDataGridReadAsync(DataGridReadDataEventArgs<ContactDto> e)
@@ -156,84 +138,40 @@ namespace IBLTermocasa.Blazor.Pages
 
         private async Task OpenCreateContactModalAsync()
         {
-            NewContact = new ContactCreateDto{
-                BirthDate = DateTime.Now,
-
-                
+            ContactInput = new ContactDto(){
             };
-            await NewContactValidations.ClearAll();
-            await CreateContactModal.Show();
-        }
-
-        private async Task CloseCreateContactModalAsync()
-        {
-            NewContact = new ContactCreateDto{
-                BirthDate = DateTime.Now,
-
-                
-            };
-            await CreateContactModal.Hide();
-        }
-
-        private async Task OpenEditContactModalAsync(ContactDto input)
-        {
-            var contact = await ContactsAppService.GetAsync(input.Id);
             
-            EditingContactId = contact.Id;
-            EditingContact = ObjectMapper.Map<ContactDto, ContactUpdateDto>(contact);
-            await EditingContactValidations.ClearAll();
-            await EditContactModal.Show();
+            SelectedContact = ContactInput;
+            NavigationManager.NavigateTo($"/contact/{Guid.Empty}");
+            //await CreateContactModal.Show();
         }
-
+        private void OpenEditContactPageAsync(ContactDto input)
+        {
+            //navigate to the page ContactDetail
+            SelectedContact = input;
+            NavigationManager.NavigateTo($"/contact/{input.Id}");
+        }
         private async Task DeleteContactAsync(ContactDto input)
         {
-            await ContactsAppService.DeleteAsync(input.Id);
-            await GetContactsAsync();
-        }
-
-        private async Task CreateContactAsync()
-        {
-            try
+            
+            bool result = await _SecureConfirmationService.ShowConfirmation(
+                "Sei sicuro di voler eliminare questo contatto?",
+                "Scrivi il nome dell contatto {0} per confermare l'eliminazione",
+                input.Name
+            );
+            if (result)
             {
-                if (await NewContactValidations.ValidateAll() == false)
-                {
-                    return;
-                }
-
-                await ContactsAppService.CreateAsync(NewContact);
+                // Procedi con la cancellazione
+                Console.WriteLine("Cancellazione in corso organization id: " + input.Id);
+                await ContactsAppService.DeleteAsync(input.Id);
                 await GetContactsAsync();
-                await CloseCreateContactModalAsync();
             }
-            catch (Exception ex)
+            else
             {
-                await HandleErrorAsync(ex);
+                // Cancellazione annullata
+                Console.WriteLine(">>>>>>>>>>>>>>>>>>>>>>>>>>>>>Cancellazione annullata...");
             }
         }
-
-        private async Task CloseEditContactModalAsync()
-        {
-            await EditContactModal.Hide();
-        }
-
-        private async Task UpdateContactAsync()
-        {
-            try
-            {
-                if (await EditingContactValidations.ValidateAll() == false)
-                {
-                    return;
-                }
-
-                await ContactsAppService.UpdateAsync(EditingContactId, EditingContact);
-                await GetContactsAsync();
-                await EditContactModal.Hide();                
-            }
-            catch (Exception ex)
-            {
-                await HandleErrorAsync(ex);
-            }
-        }
-
         private void OnSelectedCreateTabChanged(string name)
         {
             SelectedCreateTab = name;
@@ -259,6 +197,16 @@ namespace IBLTermocasa.Blazor.Pages
             Filter.Surname = surname;
             await SearchAsync();
         }
+        protected virtual async Task OnPhoneChangedAsync(string? phone)
+        {
+            Filter.PhoneInfo = phone;
+            await SearchAsync();
+        }
+        protected virtual async Task OnMailChangedAsync(string? mail)
+        {
+            Filter.MailInfo = mail;
+            await SearchAsync();
+        }
         protected virtual async Task OnConfidentialNameChangedAsync(string? confidentialName)
         {
             Filter.ConfidentialName = confidentialName;
@@ -269,33 +217,10 @@ namespace IBLTermocasa.Blazor.Pages
             Filter.JobRole = jobRole;
             await SearchAsync();
         }
-        protected virtual async Task OnMailInfoChangedAsync(string? mailInfo)
-        {
-            Filter.MailInfo = mailInfo;
-            await SearchAsync();
-        }
-        protected virtual async Task OnPhoneInfoChangedAsync(string? phoneInfo)
-        {
-            Filter.PhoneInfo = phoneInfo;
-            await SearchAsync();
-        }
-        protected virtual async Task OnAddressInfoChangedAsync(string? addressInfo)
-        {
-            Filter.AddressInfo = addressInfo;
-            await SearchAsync();
-        }
         protected virtual async Task OnTagChangedAsync(string? tag)
         {
             Filter.Tag = tag;
             await SearchAsync();
         }
-        
-
-
-
-
-
-
-
     }
 }
